// frontend/src/lib/api/generated/models/diffReviewDraftComment.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface DiffReviewDraftComment {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  body: string;
  commit_sha?: string;
  created_at: string;
  diff_head_sha?: string;
  id: string;
  line: number;
  line_type: string;
  new_line?: number;
  old_line?: number;
  old_path?: string;
  path: string;
  side: string;
  start_line?: number;
  start_side?: string;
  updated_at: string;
}