// frontend/src/lib/api/generated/models/diffReviewDraftResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { DiffReviewDraftComment } from "./diffReviewDraftComment.ts";

export interface DiffReviewDraftResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  comments: DiffReviewDraftComment[];
  draft_id?: string;
  native_multiline_ranges: boolean;
  supported_actions: string[];
}