// frontend/src/lib/api/generated/models/diffReviewLineRange.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface DiffReviewLineRange {
  commit_sha?: string;
  diff_head_sha?: string;
  line: number;
  line_type: string;
  new_line?: number;
  old_line?: number;
  old_path?: string;
  path: string;
  side: string;
  start_line?: number;
  start_side?: string;
}